adding auth

+++ Backend/src/controllers/note.ts
@@ -5,11 +5,11 @@ import mongoose from "mongoose";
 import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
-  // const authenticatedUserId = req.session.userId;
+  const authenticatedUserId = req.session.userId;
   try {
-    // // assertIsDefined(authenticatedUserId);
+    assertIsDefined(authenticatedUserId);
     const notes = await NoteModel.find({
-      //userId: authenticatedUserId.toString(),
+      userId: authenticatedUserId,
     }).exec();
     res.status(200).json(notes);
   } catch (error) {
@@ -19,9 +19,9 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const id = req.params.id;
-  // const authenticatedUserId = req.session.userId;
+  const authenticatedUserId = req.session.userId;
   try {
-    // assertIsDefined(authenticatedUserId);
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(id)) {
       throw createHttpError(400, "Unauthorized or Invalid note id");
     }
@@ -29,9 +29,9 @@ export const getNote: RequestHandler = async (req, res, next) => {
     if (!note) {
       throw createHttpError(404, "Must send valid id note not found");
     }
-    // if (!note.id.equals(authenticatedUserId)) {
-    //   throw createHttpError(401, "You cannot access this note");
-    // }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
+    }
     res.status(200).json(note);
   } catch (error) {
     next(error);
@@ -49,14 +49,14 @@ export const createNotes: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
-  // const authenticatedUserId = req.session.userId;
+  const authenticatedUserId = req.session.userId;
   try {
-    // assertIsDefined(authenticatedUserId);
+    assertIsDefined(authenticatedUserId);
     if (!title) {
       throw createHttpError(400, "Note must send title to processed");
     }
     const newNote = await NoteModel.create({
-      //userId: { authenticatedUserId },
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -82,9 +82,9 @@ export const updateNote: RequestHandler<
   const id = req.params.id;
   const newTitle = req.body.title;
   const newText = req.body.text;
-  // const authenticatedUserId = req.session.userId;
+  const authenticatedUserId = req.session.userId;
   try {
-    // assertIsDefined(authenticatedUserId);
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(id)) {
       throw createHttpError(400, "Unauthorized or Invalid note id");
     }
@@ -95,9 +95,12 @@ export const updateNote: RequestHandler<
     if (!note) {
       throw createHttpError(404, "Must send valid id note not found");
     }
-    // if (!note.id?.equals(authenticatedUserId)) {
+    // if (!note.userId?.equals(authenticatedUserId)) {
     //   throw createHttpError(401, "You cannot access this note");
     // }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
+    }
     note.title = newTitle;
     note.text = newText;
     const UpdatedNote = await note.save();
@@ -108,9 +111,9 @@ export const updateNote: RequestHandler<
 };
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const id = req.params.id;
-  // const authenticatedUserId = req.session.userId;
+  const authenticatedUserId = req.session.userId;
   try {
-    // assertIsDefined(authenticatedUserId);
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(id)) {
       throw createHttpError(400, "Unauthorized or Invalid note id");
     }
@@ -118,9 +121,9 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
     if (!note) {
       throw createHttpError(404, "Must send valid id note not found");
     }
-    // if (!note.id?.equals(authenticatedUserId)) {
-    //   throw createHttpError(401, "You cannot access this note");
-    // }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
+    }
     const deletedNote = await NoteModel.findByIdAndRemove(id).exec();
     res.status(204).json(deletedNote);
   } catch (error) {
